Small login system
Kan bare lagre en bruker atm

+++ HighFidalityModel/Scripts/Managers/AccountManager.cs
@@ -12,42 +12,67 @@ public class AccountManager : MonoBehaviour
     public TextMeshProUGUI password;
     public TextMeshProUGUI passwordVerification;
 
+    public GameObject activeWindow;
+
     public void CreateNewAccount()
     {
         if (password.text != passwordVerification.text) 
         {   
             return;
         } 
-        CreateAccount create = new CreateAccount(username.text, passwordVerification.text);
+        CreateAccount create = new CreateAccount(username.text, password.text);
+        create.SaveUser();
+        GameManager.Instance.SetActiveWindow(activeWindow);
+    }
+
+    public void LoadAccount()
+    {
+        CreateAccount create = new CreateAccount(username.text, password.text);
+
+        if (create.LoadUser())
+        {
+            GameManager.Instance.currentUser = create.user;
+            GameManager.Instance.designPorfile.userName = create.user.username;
+            GameManager.Instance.SetActiveWindow(activeWindow);
+        }
+        else Debug.Log("Wrong credentials");
     }
 }
 
 public class CreateAccount
 {
-    private User user;
+    public User user;
+    private string _userName;
+    private Hash128 _password;
+
     public CreateAccount(string userName, string password)
     {
+        _userName = userName;
+
         Hash128 hash = new Hash128();
         hash.Append(password);
 
+        _password = hash;
 
-        user = new User(userName, hash);
-        Debug.Log("User " + userName + " Has generated " + hash);
-
-        SaveUser(user);
+        user = new User(_userName, _password);
     }
 
-    private void SaveUser(User user)
+    public void SaveUser()
     {
+        User user = new User(_userName, _password);
         string json = JsonUtility.ToJson(user);
-        File.WriteAllText(Application.persistentDataPath + "/User.json", json);
-        Debug.Log("User saved");
+        if(File.Exists(Application.persistentDataPath + "/User.json")) File.WriteAllText(Application.persistentDataPath + "/User.json", json);
+        Debug.Log("User saved " + user.username);
     }
-}
 
-public class Authenticator
-{
-
+    public bool LoadUser()
+    {
+        string json = File.ReadAllText(Application.persistentDataPath + "/User.json");
+        User _user = JsonUtility.FromJson<User>(json);
+        if (_userName != _user.username) return false;
+        else if (_password != _user.password) return false;
+        else return true;
+    }
 }
 
 [Serializable]
